:recycle: Replace old `patch_predictor`
