+++ tiatoolbox/models/engine/patch_predictor.py
@@ -1,32 +1,25 @@
-"""This module implements patch level prediction."""
+"""Defines Abstract Base Class for TIAToolbox Model Engines."""
 
 from __future__ import annotations
 
-import copy
-from collections import OrderedDict
-from pathlib import Path
-from typing import TYPE_CHECKING, Callable, NoReturn
+from typing import TYPE_CHECKING
 
-import numpy as np
-import torch
-import tqdm
+from typing_extensions import Unpack
 
-import tiatoolbox.models.models_abc
-from tiatoolbox import logger
-from tiatoolbox.models.dataset.dataset_abc import PatchDataset, WSIPatchDataset
-from tiatoolbox.utils import save_as_json
-from tiatoolbox.wsicore.wsireader import VirtualWSIReader, WSIReader
+from .engine_abc import EngineABC, EngineABCRunParams
 
 if TYPE_CHECKING:  # pragma: no cover
     import os
+    from pathlib import Path
+
+    import numpy as np
+    from torch.utils.data import DataLoader
 
     from tiatoolbox.annotation import AnnotationStore
-    from tiatoolbox.typing import IntPair, Resolution, Units
+    from tiatoolbox.models.models_abc import ModelABC
+    from tiatoolbox.wsicore.wsireader import WSIReader
 
     from .io_config import ModelIOConfigABC
-
-from .engine_abc import EngineABC
-from .io_config import IOPatchPredictorConfig
 
 
 class PatchPredictor(EngineABC):
@@ -117,63 +110,141 @@ class PatchPredictor(EngineABC):
          - 0.867
 
     Args:
-        model (nn.Module):
-            Use externally defined PyTorch model for prediction with.
-            weights already loaded. Default is `None`. If provided,
-            `pretrained_model` argument is ignored.
-        pretrained_model (str):
-            Name of the existing models support by tiatoolbox for
-            processing the data. For a full list of pretrained models,
-            refer to the `docs
+        model (str | ModelABC):
+            A PyTorch model. Default is `None`.
+            The user can request pretrained models from the toolbox model zoo using
+            the list of pretrained models available at this `link
             <https://tia-toolbox.readthedocs.io/en/latest/pretrained.html>`_
             By default, the corresponding pretrained weights will also
             be downloaded. However, you can override with your own set
-            of weights via the `pretrained_weights` argument. Argument
-            is case-insensitive.
-        weights (str):
-            Path to the weight of the corresponding `pretrained_model`.
-
-          >>> predictor = PatchPredictor(
-          ...    pretrained_model="resnet18-kather100k",
-          ...    weights="resnet18_local_weight")
-
+            of weights using the `weights` parameter.
         batch_size (int):
-            Number of images fed into the model each time.
+            Number of image patches fed into the model each time in a
+            forward/backward pass. Default value is 8.
         num_loader_workers (int):
-            Number of workers to load the data. Take note that they will
-            also perform preprocessing.
+            Number of workers to load the data using :class:`torch.utils.data.Dataset`.
+            Please note that they will also perform preprocessing. Default value is 0.
+        num_post_proc_workers (int):
+            Number of workers to postprocess the results of the model.
+            Default value is 0.
+        weights (str or Path):
+            Path to the weight of the corresponding `model`.
+
+            >>> engine = EngineABC(
+            ...    model="pretrained-model",
+            ...    weights="/path/to/pretrained-local-weights.pth"
+            ... )
+
+        device (str):
+            Select the device to run the model. Please see
+            https://pytorch.org/docs/stable/tensor_attributes.html#torch.device
+            for more details on input parameters for device. Default is "cpu".
         verbose (bool):
-            Whether to output logging information.
+            Whether to output logging information. Default value is False.
 
     Attributes:
-        images (str or :obj:`pathlib.Path` or :obj:`numpy.ndarray`):
-            A HWC image or a path to WSI.
-        mode (str):
-            Type of input to process. Choose from either `patch`, `tile`
-            or `wsi`.
-        model (nn.Module):
-            Defined PyTorch model.
-        model (str):
-            Name of the existing models support by tiatoolbox for
-            processing the data. For a full list of pretrained models,
+        images (list of str or list of :obj:`Path` or NHWC :obj:`numpy.ndarray`):
+            A list of image patches in NHWC format as a numpy array
+            or a list of str/paths to WSIs.
+        masks (list of str or list of :obj:`Path` or NHWC :obj:`numpy.ndarray`):
+            A list of tissue masks or binary masks corresponding to processing area of
+            input images. These can be a list of numpy arrays or paths to
+            the saved image masks. These are only utilized when patch_mode is False.
+            Patches are only generated within a masked area.
+            If not provided, then a tissue mask will be automatically
+            generated for whole slide images.
+        patch_mode (str):
+            Whether to treat input images as a set of image patches. TIAToolbox defines
+            an image as a patch if HWC of the input image matches with the HWC expected
+            by the model. If HWC of the input image does not match with the HWC expected
+            by the model, then the patch_mode must be set to False which will allow the
+            engine to extract patches from the input image.
+            In this case, when the patch_mode is False the input images are treated
+            as WSIs. Default value is True.
+        model (str | ModelABC):
+            A PyTorch model or a name of an existing model from the TIAToolbox model zoo
+            for processing the data. For a full list of pretrained models,
             refer to the `docs
             <https://tia-toolbox.readthedocs.io/en/latest/pretrained.html>`_
             By default, the corresponding pretrained weights will also
             be downloaded. However, you can override with your own set
-            of weights via the `pretrained_weights` argument. Argument
-            is case insensitive.
+            of weights via the `weights` argument. Argument
+            is case-insensitive.
+        ioconfig (ModelIOConfigABC):
+            Input IO configuration of type :class:`ModelIOConfigABC` to run the Engine.
+        _ioconfig (ModelIOConfigABC):
+            Runtime ioconfig.
+        return_labels (bool):
+            Whether to return the labels with the predictions.
+        merge_predictions (bool):
+            Whether to merge the predictions to form a 2-dimensional
+            map. This is only applicable if `patch_mode` is False in inference.
+            Default is False.
+        resolution (Resolution):
+            Resolution used for reading the image. Please see
+            :obj:`WSIReader` for details.
+        units (Units):
+            Units of resolution used for reading the image. Choose
+            from either `level`, `power` or `mpp`. Please see
+            :obj:`WSIReader` for details.
+        patch_input_shape (tuple):
+            Shape of patches input to the model as tupled of HW. Patches are at
+            requested read resolution, not with respect to level 0,
+            and must be positive.
+        stride_shape (tuple):
+            Stride used during WSI processing. Stride is
+            at requested read resolution, not with respect to
+            level 0, and must be positive. If not provided,
+            `stride_shape=patch_input_shape`.
         batch_size (int):
             Number of images fed into the model each time.
-        num_loader_worker (int):
-            Number of workers used in torch.utils.data.DataLoader.
+        cache_mode (bool):
+            Whether to run the Engine in cache_mode. For large datasets,
+            we recommend to set this to True to avoid out of memory errors.
+            For smaller datasets, the cache_mode is set to False as
+            the results can be saved in memory. cache_mode is always True when
+            processing WSIs i.e., when `patch_mode` is False. Default value is False.
+        cache_size (int):
+            Specifies how many image patches to process in a batch when
+            cache_mode is set to True. If cache_size is less than the batch_size
+            batch_size is set to cache_size. Default value is 10,000.
+        labels (list | None):
+                List of labels. Only a single label per image is supported.
+        device (str):
+            :class:`torch.device` to run the model.
+            Select the device to run the model. Please see
+            https://pytorch.org/docs/stable/tensor_attributes.html#torch.device
+            for more details on input parameters for device. Default value is "cpu".
+        num_loader_workers (int):
+            Number of workers used in :class:`torch.utils.data.DataLoader`.
+        num_post_proc_workers (int):
+            Number of workers to postprocess the results of the model.
+        return_labels (bool):
+            Whether to return the output labels. Default value is False.
+        merge_predictions (bool):
+            Whether to merge WSI predictions into a single file. Default value is False.
+        resolution (Resolution):
+            Resolution used for reading the image. Please see
+            :class:`WSIReader` for details.
+            When `patch_mode` is True, the input image patches are expected to be at
+            the correct resolution and units. When `patch_mode` is False, the patches
+            are extracted at the requested resolution and units. Default value is 1.0.
+        units (Units):
+            Units of resolution used for reading the image. Choose
+            from either `baseline`, `level`, `power` or `mpp`. Please see
+            :class:`WSIReader` for details.
+            When `patch_mode` is True, the input image patches are expected to be at
+            the correct resolution and units. When `patch_mode` is False, the patches
+            are extracted at the requested resolution and units.
+            Default value is `baseline`.
         verbose (bool):
-            Whether to output logging information.
+            Whether to output logging information. Default value is False.
 
     Examples:
         >>> # list of 2 image patches as input
         >>> data = ['path/img.svs', 'path/img.svs']
         >>> predictor = PatchPredictor(pretrained_model="resnet18-kather100k")
-        >>> output = predictor.predict(data, mode='patch')
+        >>> output = predictor.run(data, mode='patch')
 
         >>> # array of list of 2 image patches as input
         >>> data = np.array([img1, img2])
@@ -183,17 +254,17 @@ class PatchPredictor(EngineABC):
         >>> # list of 2 image patch files as input
         >>> data = ['path/img.png', 'path/img.png']
         >>> predictor = PatchPredictor(pretrained_model="resnet18-kather100k")
-        >>> output = predictor.predict(data, mode='patch')
+        >>> output = predictor.run(data, mode='patch')
 
         >>> # list of 2 image tile files as input
         >>> tile_file = ['path/tile1.png', 'path/tile2.png']
         >>> predictor = PatchPredictor(pretraind_model="resnet18-kather100k")
-        >>> output = predictor.predict(tile_file, mode='tile')
+        >>> output = predictor.run(tile_file, mode='tile')
 
         >>> # list of 2 wsi files as input
         >>> wsi_file = ['path/wsi1.svs', 'path/wsi2.svs']
         >>> predictor = PatchPredictor(pretraind_model="resnet18-kather100k")
-        >>> output = predictor.predict(wsi_file, mode='wsi')
+        >>> output = predictor.run(wsi_file, mode='wsi')
 
     References:
         [1] Kather, Jakob Nikolas, et al. "Predicting survival from colorectal cancer
@@ -208,526 +279,143 @@ class PatchPredictor(EngineABC):
 
     def __init__(
         self: PatchPredictor,
+        model: str | ModelABC,
         batch_size: int = 8,
         num_loader_workers: int = 0,
         num_post_proc_workers: int = 0,
-        model: torch.nn.Module = None,
-        pretrained_model: str | None = None,
-        weights: str | None = None,
+        weights: str | Path | None = None,
         *,
+        device: str = "cpu",
         verbose: bool = True,
     ) -> None:
         """Initialize :class:`PatchPredictor`."""
         super().__init__(
+            model=model,
             batch_size=batch_size,
             num_loader_workers=num_loader_workers,
             num_post_proc_workers=num_post_proc_workers,
-            model=model,
-            pretrained_model=pretrained_model,
             weights=weights,
+            device=device,
             verbose=verbose,
         )
 
-    def pre_process_wsi(self: PatchPredictor) -> NoReturn:
-        """Pre-process a WSI."""
-
-    def infer_wsi(self: PatchPredictor) -> NoReturn:
-        """Model inference on a WSI."""
-
-    def save_predictions(
+    def get_dataloader(
         self: PatchPredictor,
-        raw_predictions: dict,
+        images: Path,
+        masks: Path | None = None,
+        labels: list | None = None,
+        ioconfig: ModelIOConfigABC | None = None,
+        *,
+        patch_mode: bool = True,
+    ) -> DataLoader:
+        """Pre-process images and masks and return dataloader for inference.
+
+        Args:
+            images (list of str or :class:`Path` or :class:`numpy.ndarray`):
+                A list of image patches in NHWC format as a numpy array
+                or a list of str/paths to WSIs. When `patch_mode` is False
+                the function expects list of str/paths to WSIs.
+            masks (list | None):
+                List of masks. Only utilised when patch_mode is False.
+                Patches are only generated within a masked area.
+                If not provided, then a tissue mask will be automatically
+                generated for whole slide images.
+            labels (list | None):
+                List of labels. Only a single label per image is supported.
+            ioconfig (ModelIOConfigABC):
+                A :class:`ModelIOConfigABC` object.
+            patch_mode (bool):
+                Whether to treat input image as a patch or WSI.
+
+        Returns:
+            DataLoader:
+                :class:`DataLoader` for inference.
+
+
+        """
+        return super().get_dataloader(
+            images,
+            masks,
+            labels,
+            ioconfig,
+            patch_mode=patch_mode,
+        )
+
+    def infer_wsi(
+        self: EngineABC,
+        dataloader: DataLoader,
+        save_path: Path,
+        **kwargs: EngineABCRunParams,
+    ) -> Path:
+        """Model inference on a WSI.
+
+        Args:
+            dataloader (DataLoader):
+                A torch dataloader to process WSIs.
+
+            save_path (Path):
+                Path to save the intermediate output. The intermediate output is saved
+                in a zarr file.
+            **kwargs (EngineABCRunParams):
+                Keyword Args to update setup_patch_dataset() method attributes. See
+                :class:`EngineRunParams` for accepted keyword arguments.
+
+        Returns:
+            save_path (Path):
+                Path to zarr file where intermediate output is saved.
+
+        """
+        _ = kwargs.get("patch_mode", False)
+        return self.infer_patches(
+            dataloader=dataloader,
+            save_path=save_path,
+            return_coordinates=True,
+        )
+
+    def post_process_wsi(
+        self: EngineABC,
+        raw_predictions: dict | Path,
+        **kwargs: Unpack[EngineABCRunParams],
+    ) -> dict | Path:
+        """Post process WSI output.
+
+        Takes the raw output from patch predictions and post-process to improve the
+        results e.g., using information from neighbouring patches.
+
+        """
+        return super().post_process_wsi(
+            raw_predictions=raw_predictions,
+            **kwargs,
+        )
+
+    def save_wsi_output(
+        self: EngineABC,
+        processed_output: Path,
         output_type: str,
-    ) -> None:
-        """Post-process an image patch."""
-
-    def save_wsi_output(self: PatchPredictor) -> NoReturn:
-        """Post-process a WSI."""
-
-    @staticmethod
-    def merge_predictions(
-        img: str | Path | np.ndarray,
-        output: dict,
-        resolution: Resolution | None = None,
-        units: Units | None = None,
-        post_proc_func: Callable | None = None,
-        *,
-        return_raw: bool = False,
-    ) -> np.ndarray:
-        """Merge patch level predictions to form a 2-dimensional prediction map.
-
-        #! Improve how the below reads.
-        The prediction map will contain values from 0 to N, where N is
-        the number of classes. Here, 0 is the background which has not
-        been processed by the model and N is the number of classes
-        predicted by the model.
+        **kwargs: Unpack[EngineABCRunParams],
+    ) -> Path:
+        """Aggregate the output at the WSI level and save to file.
 
         Args:
-            img (:obj:`str` or :obj:`pathlib.Path` or :class:`numpy.ndarray`):
-              A HWC image or a path to WSI.
-            output (dict):
-                Output generated by the model.
-            resolution (Resolution):
-                Resolution of merged predictions.
-            units (Units):
-                Units of resolution used when merging predictions. This
-                must be the same `units` used when processing the data.
-            post_proc_func (callable):
-                A function to post-process raw prediction from model. By
-                default, internal code uses the `np.argmax` function.
-            return_raw (bool):
-                Return raw result without applying the `postproc_func`
-                on the assembled image.
+            processed_output (Path):
+                Path to Zarr file with intermediate results.
+            output_type (str):
+                The desired output type for resulting patch dataset.
+            **kwargs (EngineABCRunParams):
+                Keyword Args to update setup_patch_dataset() method attributes.
 
-        Returns:
-            :class:`numpy.ndarray`:
-                Merged predictions as a 2D array.
-
-        Examples:
-            >>> # pseudo output dict from model with 2 patches
-            >>> output = {
-            ...     'resolution': 1.0,
-            ...     'units': 'baseline',
-            ...     'probabilities': [[0.45, 0.55], [0.90, 0.10]],
-            ...     'predictions': [1, 0],
-            ...     'coordinates': [[0, 0, 2, 2], [2, 2, 4, 4]],
-            ... }
-            >>> merged = PatchPredictor.merge_predictions(
-            ...         np.zeros([4, 4]),
-            ...         output,
-            ...         resolution=1.0,
-            ...         units='baseline'
-            ... )
-            >>> merged
-            ... array([[2, 2, 0, 0],
-            ...    [2, 2, 0, 0],
-            ...    [0, 0, 1, 1],
-            ...    [0, 0, 1, 1]])
+        Returns: (AnnotationStore or Path):
+            If the output_type is "AnnotationStore", the function returns the patch
+            predictor output as an SQLiteStore containing Annotations stored in a `.db`
+            file. Otherwise, the function defaults to returning patch predictor output
+            stored in a `.zarr` file.
 
         """
-        reader = WSIReader.open(img)
-        if isinstance(reader, VirtualWSIReader):
-            logger.warning(
-                "Image is not pyramidal hence read is forced to be "
-                "at `units='baseline'` and `resolution=1.0`.",
-                stacklevel=2,
-            )
-            resolution = 1.0
-            units = "baseline"
-
-        canvas_shape = reader.slide_dimensions(resolution=resolution, units=units)
-        canvas_shape = canvas_shape[::-1]  # XY to YX
-
-        # may crash here, do we need to deal with this ?
-        output_shape = reader.slide_dimensions(
-            resolution=output["resolution"],
-            units=output["units"],
+        return super().save_wsi_output(
+            processed_output=processed_output,
+            output_type=output_type,
+            **kwargs,
         )
-        output_shape = output_shape[::-1]  # XY to YX
-        fx = np.array(canvas_shape) / np.array(output_shape)
-
-        if "probabilities" not in output:
-            coordinates = output["coordinates"]
-            predictions = output["predictions"]
-            denominator = None
-            output = np.zeros(list(canvas_shape), dtype=np.float32)
-        else:
-            coordinates = output["coordinates"]
-            predictions = output["probabilities"]
-            num_class = np.array(predictions[0]).shape[0]
-            denominator = np.zeros(canvas_shape)
-            output = np.zeros([*list(canvas_shape), num_class], dtype=np.float32)
-
-        for idx, bound in enumerate(coordinates):
-            prediction = predictions[idx]
-            # assumed to be in XY
-            # top-left for output placement
-            tl = np.ceil(np.array(bound[:2]) * fx).astype(np.int32)
-            # bot-right for output placement
-            br = np.ceil(np.array(bound[2:]) * fx).astype(np.int32)
-            output[tl[1] : br[1], tl[0] : br[0]] += prediction
-            if denominator is not None:
-                denominator[tl[1] : br[1], tl[0] : br[0]] += 1
-
-        # deal with overlapping regions
-        if denominator is not None:
-            output = output / (np.expand_dims(denominator, -1) + 1.0e-8)
-            if not return_raw:
-                # convert raw probabilities to predictions
-                if post_proc_func is not None:
-                    output = post_proc_func(output)
-                else:
-                    output = np.argmax(output, axis=-1)
-                # to make sure background is 0 while class will be 1...N
-                output[denominator > 0] += 1
-        return output
-
-    def _predict_engine(
-        self: PatchPredictor,
-        dataset: torch.utils.data.Dataset,
-        *,
-        return_probabilities: bool = False,
-        return_labels: bool = False,
-        return_coordinates: bool = False,
-        device: str = "cpu",
-    ) -> np.ndarray:
-        """Make a prediction on a dataset. The dataset may be mutated.
-
-        Args:
-            dataset (torch.utils.data.Dataset):
-                PyTorch dataset object created using
-                `tiatoolbox.models.data.classification.Patch_Dataset`.
-            return_probabilities (bool):
-                Whether to return per-class probabilities.
-            return_labels (bool):
-                Whether to return labels.
-            return_coordinates (bool):
-                Whether to return patch coordinates.
-            device (str):
-                Select the device to run the model. Default is "cpu".
-
-        Returns:
-            :class:`numpy.ndarray`:
-                Model predictions of the input dataset
-
-        """
-        dataset.preproc_func = self.model.preproc_func
-
-        # preprocessing must be defined with the dataset
-        dataloader = torch.utils.data.DataLoader(
-            dataset,
-            num_workers=self.num_loader_workers,
-            batch_size=self.batch_size,
-            drop_last=False,
-            shuffle=False,
-        )
-
-        if self.verbose:
-            pbar = tqdm.tqdm(
-                total=int(len(dataloader)),
-                leave=True,
-                ncols=80,
-                ascii=True,
-                position=0,
-            )
-
-        # use external for testing
-        model = tiatoolbox.models.models_abc.model_to(model=self.model, device=device)
-
-        cum_output = {
-            "probabilities": [],
-            "predictions": [],
-            "coordinates": [],
-            "labels": [],
-        }
-        for _, batch_data in enumerate(dataloader):
-            batch_output_probabilities = self.model.infer_batch(
-                model,
-                batch_data["image"],
-                device=device,
-            )
-            # We get the index of the class with the maximum probability
-            batch_output_predictions = self.model.postproc_func(
-                batch_output_probabilities,
-            )
-
-            # tolist might be very expensive
-            cum_output["probabilities"].extend(batch_output_probabilities.tolist())
-            cum_output["predictions"].extend(batch_output_predictions.tolist())
-            if return_coordinates:
-                cum_output["coordinates"].extend(batch_data["coords"].tolist())
-            if return_labels:  # be careful of `s`
-                # We do not use tolist here because label may be of mixed types
-                # and hence collated as list by torch
-                cum_output["labels"].extend(list(batch_data["label"]))
-
-            if self.verbose:
-                pbar.update()
-        if self.verbose:
-            pbar.close()
-
-        if not return_probabilities:
-            cum_output.pop("probabilities")
-        if not return_labels:
-            cum_output.pop("labels")
-        if not return_coordinates:
-            cum_output.pop("coordinates")
-
-        return cum_output
-
-    def _update_ioconfig(
-        self: PatchPredictor,
-        ioconfig: IOPatchPredictorConfig,
-        patch_input_shape: IntPair,
-        stride_shape: IntPair,
-        resolution: Resolution,
-        units: Units,
-    ) -> IOPatchPredictorConfig:
-        """Update the ioconfig.
-
-        Args:
-            ioconfig (:class:`IOPatchPredictorConfig`):
-                Input ioconfig for PatchPredictor.
-            patch_input_shape (tuple):
-                Size of patches input to the model. Patches are at
-                requested read resolution, not with respect to level 0,
-                and must be positive.
-            stride_shape (tuple):
-                Stride using during tile and WSI processing. Stride is
-                at requested read resolution, not with respect to
-                level 0, and must be positive. If not provided,
-                `stride_shape=patch_input_shape`.
-            resolution (Resolution):
-                Resolution used for reading the image. Please see
-                :obj:`WSIReader` for details.
-            units (Units):
-                Units of resolution used for reading the image.
-
-        Returns:
-            Updated Patch Predictor IO configuration.
-
-        """
-        config_flag = (
-            patch_input_shape is None,
-            resolution is None,
-            units is None,
-        )
-        if ioconfig:
-            return ioconfig
-
-        if self.ioconfig is None and any(config_flag):
-            msg = (
-                "Must provide either "
-                "`ioconfig` or `patch_input_shape`, `resolution`, and `units`."
-            )
-            raise ValueError(
-                msg,
-            )
-
-        if stride_shape is None:
-            stride_shape = patch_input_shape
-
-        if self.ioconfig:
-            ioconfig = copy.deepcopy(self.ioconfig)
-            # ! not sure if there is a nicer way to set this
-            if patch_input_shape is not None:
-                ioconfig.patch_input_shape = patch_input_shape
-            if stride_shape is not None:
-                ioconfig.stride_shape = stride_shape
-            if resolution is not None:
-                ioconfig.input_resolutions[0]["resolution"] = resolution
-            if units is not None:
-                ioconfig.input_resolutions[0]["units"] = units
-
-            return ioconfig
-
-        return IOPatchPredictorConfig(
-            input_resolutions=[{"resolution": resolution, "units": units}],
-            patch_input_shape=patch_input_shape,
-            stride_shape=stride_shape,
-            output_resolutions=[],
-        )
-
-    def _predict_patch(
-        self: PatchPredictor,
-        imgs: list | np.ndarray,
-        labels: list,
-        *,
-        return_probabilities: bool,
-        return_labels: bool,
-        device: str,
-    ) -> np.ndarray:
-        """Process patch mode.
-
-        Args:
-            imgs (list, ndarray):
-                List of inputs to process. when using `patch` mode, the
-                input must be either a list of images, a list of image
-                file paths or a numpy array of an image list. When using
-                `tile` or `wsi` mode, the input must be a list of file
-                paths.
-            labels:
-                List of labels. If using `tile` or `wsi` mode, then only
-                a single label per image tile or whole-slide image is
-                supported.
-            return_probabilities (bool):
-                Whether to return per-class probabilities.
-            return_labels (bool):
-                Whether to return the labels with the predictions.
-            device (str):
-                Select the device to run the engine.
-
-        Returns:
-            :class:`numpy.ndarray`:
-                Model predictions of the input dataset
-
-        """
-        if labels:
-            # if a labels is provided, then return with the prediction
-            return_labels = bool(labels)
-
-        if labels and len(labels) != len(imgs):
-            msg = f"len(labels) != len(imgs) : {len(labels)} != {len(imgs)}"
-            raise ValueError(
-                msg,
-            )
-
-        # don't return coordinates if patches are already extracted
-        return_coordinates = False
-        dataset = PatchDataset(imgs, labels)
-        return self._predict_engine(
-            dataset,
-            return_probabilities=return_probabilities,
-            return_labels=return_labels,
-            return_coordinates=return_coordinates,
-            device=device,
-        )
-
-    def _predict_tile_wsi(  # noqa: PLR0913
-        self: PatchPredictor,
-        imgs: list,
-        masks: list | None,
-        labels: list,
-        mode: str,
-        ioconfig: IOPatchPredictorConfig,
-        save_dir: str | Path,
-        highest_input_resolution: list[dict],
-        *,
-        save_output: bool,
-        return_probabilities: bool,
-        merge_predictions: bool,
-        on_gpu: bool,
-    ) -> list | dict:
-        """Predict on Tile and WSIs.
-
-        Args:
-            imgs (list, ndarray):
-                List of inputs to process. when using `patch` mode, the
-                input must be either a list of images, a list of image
-                file paths or a numpy array of an image list. When using
-                `tile` or `wsi` mode, the input must be a list of file
-                paths.
-            masks (list or None):
-                List of masks. Only utilised when processing image tiles
-                and whole-slide images. Patches are only processed if
-                they are within a masked area. If not provided, then a
-                tissue mask will be automatically generated for
-                whole-slide images or the entire image is processed for
-                image tiles.
-            labels:
-                List of labels. If using `tile` or `wsi` mode, then only
-                a single label per image tile or whole-slide image is
-                supported.
-            mode (str):
-                Type of input to process. Choose from either `patch`,
-                `tile` or `wsi`.
-            return_probabilities (bool):
-                Whether to return per-class probabilities.
-            on_gpu (bool):
-                Whether to run model on the GPU.
-            ioconfig (IOPatchPredictorConfig):
-                Patch Predictor IO configuration..
-            merge_predictions (bool):
-                Whether to merge the predictions to form a 2-dimensional
-                map. This is only applicable for `mode='wsi'` or
-                `mode='tile'`.
-            save_dir (str or pathlib.Path):
-                Output directory when processing multiple tiles and
-                whole-slide images. By default, it is folder `output`
-                where the running script is invoked.
-            save_output (bool):
-                Whether to save output for a single file. default=False
-            highest_input_resolution (list(dict)):
-                Highest available input resolution.
-
-
-        Returns:
-            dict:
-                Results are saved to `save_dir` and a dictionary indicating save
-                location for each input is returned. The dict is in the following
-                format:
-                    - img_path: path of the input image.
-                        - raw: path to save location for raw prediction,
-                          saved in .json.
-                        - merged: path to .npy contain merged
-                          predictions if
-                        `merge_predictions` is `True`.
-
-        """
-        # return coordinates of patches processed within a tile / whole-slide image
-        return_coordinates = True
-
-        input_is_path_like = isinstance(imgs[0], (str, Path))
-        default_save_dir = (
-            imgs[0].parent / "output" if input_is_path_like else Path.cwd()
-        )
-        save_dir = default_save_dir if save_dir is None else Path(save_dir)
-
-        # None if no output
-        outputs = None
-
-        self._ioconfig = ioconfig
-        # generate a list of output file paths if number of input images > 1
-        file_dict = OrderedDict()
-
-        if len(imgs) > 1:
-            save_output = True
-
-        for idx, img_path in enumerate(imgs):
-            img_path_ = Path(img_path)
-            img_label = None if labels is None else labels[idx]
-            img_mask = None if masks is None else masks[idx]
-
-            dataset = WSIPatchDataset(
-                img_path_,
-                mode=mode,
-                mask_path=img_mask,
-                patch_input_shape=ioconfig.patch_input_shape,
-                stride_shape=ioconfig.stride_shape,
-                resolution=ioconfig.input_resolutions[0]["resolution"],
-                units=ioconfig.input_resolutions[0]["units"],
-            )
-            output_model = self._predict_engine(
-                dataset,
-                return_labels=False,
-                return_probabilities=return_probabilities,
-                return_coordinates=return_coordinates,
-                on_gpu=on_gpu,
-            )
-            output_model["label"] = img_label
-            # add extra information useful for downstream analysis
-            output_model["pretrained_model"] = self.model
-            output_model["resolution"] = highest_input_resolution["resolution"]
-            output_model["units"] = highest_input_resolution["units"]
-
-            outputs = [output_model]  # assign to a list
-            merged_prediction = None
-            if merge_predictions:
-                merged_prediction = self.merge_predictions(
-                    img_path_,
-                    output_model,
-                    resolution=output_model["resolution"],
-                    units=output_model["units"],
-                    post_proc_func=self.model.postproc,
-                )
-                outputs.append(merged_prediction)
-
-            if save_output:
-                # dynamic 0 padding
-                img_code = f"{idx:0{len(str(len(imgs)))}d}"
-
-                save_info = {}
-                save_path = save_dir / img_code
-                raw_save_path = f"{save_path}.raw.json"
-                save_info["raw"] = raw_save_path
-                save_as_json(output_model, raw_save_path)
-                if merge_predictions:
-                    merged_file_path = f"{save_path}.merged.npy"
-                    np.save(merged_file_path, merged_prediction)
-                    save_info["merged"] = merged_file_path
-                file_dict[str(img_path_)] = save_info
-
-        return file_dict if save_output else outputs
 
     def run(
         self: EngineABC,
@@ -740,102 +428,51 @@ class PatchPredictor(EngineABC):
         save_dir: os | Path | None = None,  # None will not save output
         overwrite: bool = False,
         output_type: str = "dict",
-        **kwargs: dict,
-    ) -> AnnotationStore | str:
-        """Run engine."""
-        super().run(
-            images=images,
-            masks=masks,
-            labels=labels,
-            ioconfig=ioconfig,
-            patch_mode=patch_mode,
-            save_dir=save_dir,
-            overwrite=overwrite,
-            output_type=output_type,
-            **kwargs,
-        )
-
-    def predict(  # noqa: PLR0913
-        self: PatchPredictor,
-        imgs: list,
-        masks: list | None = None,
-        labels: list | None = None,
-        mode: str = "patch",
-        ioconfig: IOPatchPredictorConfig | None = None,
-        patch_input_shape: tuple[int, int] | None = None,
-        stride_shape: tuple[int, int] | None = None,
-        resolution: Resolution | None = None,
-        units: Units = None,
-        *,
-        return_probabilities: bool = False,
-        return_labels: bool = False,
-        on_gpu: bool = True,
-        merge_predictions: bool = False,
-        save_dir: str | Path | None = None,
-        save_output: bool = False,
-    ) -> np.ndarray | list | dict:
-        """Make a prediction for a list of input data.
+        **kwargs: Unpack[EngineABCRunParams],
+    ) -> AnnotationStore | Path | str | dict:
+        """Run the engine on input images.
 
         Args:
-            imgs (list, ndarray):
+            images (list, ndarray):
                 List of inputs to process. when using `patch` mode, the
                 input must be either a list of images, a list of image
-                file paths or a numpy array of an image list. When using
-                `tile` or `wsi` mode, the input must be a list of file
-                paths.
-            masks (list):
-                List of masks. Only utilised when processing image tiles
-                and whole-slide images. Patches are only processed if
-                they are within a masked area. If not provided, then a
-                tissue mask will be automatically generated for
-                whole-slide images or the entire image is processed for
-                image tiles.
-            labels:
-                List of labels. If using `tile` or `wsi` mode, then only
-                a single label per image tile or whole-slide image is
-                supported.
-            mode (str):
-                Type of input to process. Choose from either `patch`,
-                `tile` or `wsi`.
-            return_probabilities (bool):
-                Whether to return per-class probabilities.
-            return_labels (bool):
-                Whether to return the labels with the predictions.
-            on_gpu (bool):
-                Whether to run model on the GPU.
+                file paths or a numpy array of an image list.
+            masks (list | None):
+                List of masks. Only utilised when patch_mode is False.
+                Patches are only generated within a masked area.
+                If not provided, then a tissue mask will be automatically
+                generated for whole slide images.
+            labels (list | None):
+                List of labels. Only a single label per image is supported.
+            patch_mode (bool):
+                Whether to treat input image as a patch or WSI.
+                default = True.
             ioconfig (IOPatchPredictorConfig):
-                Patch Predictor IO configuration.
-            patch_input_shape (tuple):
-                Size of patches input to the model. Patches are at
-                requested read resolution, not with respect to level 0,
-                and must be positive.
-            stride_shape (tuple):
-                Stride using during tile and WSI processing. Stride is
-                at requested read resolution, not with respect to
-                level 0, and must be positive. If not provided,
-                `stride_shape=patch_input_shape`.
-            resolution (Resolution):
-                Resolution used for reading the image. Please see
-                :obj:`WSIReader` for details.
-            units (Units):
-                Units of resolution used for reading the image. Choose
-                from either `level`, `power` or `mpp`. Please see
-                :obj:`WSIReader` for details.
-            merge_predictions (bool):
-                Whether to merge the predictions to form a 2-dimensional
-                map. This is only applicable for `mode='wsi'` or
-                `mode='tile'`.
+                IO configuration.
             save_dir (str or pathlib.Path):
-                Output directory when processing multiple tiles and
-                whole-slide images. By default, it is folder `output`
-                where the running script is invoked.
-            save_output (bool):
-                Whether to save output for a single file. default=False
+                Output directory to save the results.
+                If save_dir is not provided when patch_mode is False,
+                then for a single image the output is created in the current directory.
+                If there are multiple WSIs as input then the user must provide
+                path to save directory otherwise an OSError will be raised.
+            overwrite (bool):
+                Whether to overwrite the results. Default = False.
+            output_type (str):
+                The format of the output type. "output_type" can be
+                "zarr" or "AnnotationStore". Default value is "zarr".
+                When saving in the zarr format the output is saved using the
+                `python zarr library <https://zarr.readthedocs.io/en/stable/>`__
+                as a zarr group. If the required output type is an "AnnotationStore"
+                then the output will be intermediately saved as zarr but converted
+                to :class:`AnnotationStore` and saved as a `.db` file
+                at the end of the loop.
+            **kwargs (EngineABCRunParams):
+                Keyword Args to update :class:`EngineABC` attributes during runtime.
 
         Returns:
             (:class:`numpy.ndarray`, dict):
                 Model predictions of the input dataset. If multiple
-                image tiles or whole-slide images are provided as input,
+                whole slide images are provided as input,
                 or save_output is True, then results are saved to
                 `save_dir` and a dictionary indicating save location for
                 each input is returned.
@@ -850,79 +487,34 @@ class PatchPredictor(EngineABC):
 
         Examples:
             >>> wsis = ['wsi1.svs', 'wsi2.svs']
-            >>> predictor = PatchPredictor(
-            ...                 pretrained_model="resnet18-kather100k")
-            >>> output = predictor.predict(wsis, mode="wsi")
+            >>> class PatchPredictor(EngineABC):
+            >>> # Define all Abstract methods.
+            >>>     ...
+            >>> predictor = PatchPredictor(model="resnet18-kather100k")
+            >>> output = predictor.run(image_patches, patch_mode=True)
+            >>> output
+            ... "/path/to/Output.db"
+            >>> output = predictor.run(
+            >>>     image_patches,
+            >>>     patch_mode=True,
+            >>>     output_type="zarr")
+            >>> output
+            ... "/path/to/Output.zarr"
+            >>> output = predictor.run(wsis, patch_mode=False)
             >>> output.keys()
             ... ['wsi1.svs', 'wsi2.svs']
             >>> output['wsi1.svs']
-            ... {'raw': '0.raw.json', 'merged': '0.merged.npy'}
-            >>> output['wsi2.svs']
-            ... {'raw': '1.raw.json', 'merged': '1.merged.npy'}
+            ... {'/path/to/wsi1.db'}
 
         """
-        if mode not in ["patch", "wsi", "tile"]:
-            msg = f"{mode} is not a valid mode. Use either `patch`, `tile` or `wsi`"
-            raise ValueError(
-                msg,
-            )
-        if mode == "patch":
-            return self._predict_patch(
-                imgs,
-                labels,
-                return_probabilities,
-                return_labels,
-                on_gpu,
-            )
-
-        if not isinstance(imgs, list):
-            msg = "Input to `tile` and `wsi` mode must be a list of file paths."
-            raise TypeError(
-                msg,
-            )
-
-        if mode == "wsi" and masks is not None and len(masks) != len(imgs):
-            msg = f"len(masks) != len(imgs) : {len(masks)} != {len(imgs)}"
-            raise ValueError(
-                msg,
-            )
-
-        ioconfig = self._update_ioconfig(
-            ioconfig,
-            patch_input_shape,
-            stride_shape,
-            resolution,
-            units,
-        )
-        if mode == "tile":
-            logger.warning(
-                "WSIPatchDataset only reads image tile at "
-                '`units="baseline"`. Resolutions will be converted '
-                "to baseline value.",
-                stacklevel=2,
-            )
-            ioconfig = ioconfig.to_baseline()
-
-        fx_list = ioconfig.scale_to_highest(
-            ioconfig.input_resolutions,
-            ioconfig.input_resolutions[0]["units"],
-        )
-        fx_list = zip(fx_list, ioconfig.input_resolutions)
-        fx_list = sorted(fx_list, key=lambda x: x[0])
-        highest_input_resolution = fx_list[0][1]
-
-        save_dir = self._prepare_save_dir(save_dir, imgs)
-
-        return self._predict_tile_wsi(
-            imgs,
-            masks,
-            labels,
-            mode,
-            return_probabilities,
-            on_gpu,
-            ioconfig,
-            merge_predictions,
-            save_dir,
-            save_output,
-            highest_input_resolution,
+        return super().run(
+            images=images,
+            masks=masks,
+            labels=labels,
+            ioconfig=ioconfig,
+            patch_mode=patch_mode,
+            save_dir=save_dir,
+            overwrite=overwrite,
+            output_type=output_type,
+            **kwargs,
         )
